+++ provider-node/src/main.rs
@@ -1,4 +1,4 @@
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
-    storage_provider_node::cli::run().await
+    storage_provider_node::command::run().await
 }
